requirements.txt
Babel==2.5.3
certifi==2018.1.18
chardet==3.0.4
defusedxml==0.5.0
diff-match-patch==20121119
dj-database-url==0.4.2
dj-static==0.0.6
Django==2.0
django-allauth==0.34.0
django-bootstrap-form==3.3
django-crispy-forms==1.7.0
django-disqus==0.5
django-filter==1.1.0
django-import-export==0.6.1
django-markdown==0.8.4
django-nocaptcha-recaptcha==0.0.19
django-phonenumber-field==2.0.0
django-robots==3.0
django-suit==0.2.25
django-toolbelt==0.0.1
django-wysiwyg-redactor==0.5.1
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
gunicorn==19.7.1
idna==2.6
jdcal==1.3
Markdown==2.6.10
markdown2==2.3.5
mock==2.0.0
oauthlib==2.0.6
odfpy==1.3.6
olefile==0.44
openpyxl==2.4.9
pbr==3.1.1
phonenumberslite==8.8.10
Pillow==4.3.0
psycopg2==2.7.3.2
Pygments==2.2.0
PyJWT==1.5.3
python3-openid==3.1.0
pytz==2017.3
PyYAML==3.12
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
static3==0.7.0
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.22
xlrd==1.1.0
xlwt==1.3.0
